src/requirements.txt
streamlit==1.22.0
opencv-python==4.7.0.72
pillow==9.5.0
numpy==1.24.3
pandas
streamlit-aggrid
pymongo==4.3.3
